core, eth, internal, cmd: rework EVM constructor (#30745)
This pull request refactors the EVM constructor by removing the
TxContext parameter.

The EVM object is frequently overused. Ideally, only a single EVM
instance should be created and reused throughout the entire state
transition of a block, with the transaction context switched as needed
by calling evm.SetTxContext.

Unfortunately, in some parts of the code, the EVM object is repeatedly
created, resulting in unnecessary complexity. This pull request is the
first step towards gradually improving and simplifying this setup.

---------

Co-authored-by: Martin Holst Swende <martin@swende.se>
# Conflicts:
#	cmd/evm/internal/t8ntool/execution.go
#	core/chain_makers.go
#	core/state_prefetcher.go
#	core/state_processor.go
#	core/verkle_witness_test.go
#	core/vm/evm.go
#	core/vm/gas_table_test.go
#	core/vm/instructions_test.go
#	eth/state_accessor.go
#	eth/tracers/api.go
#	internal/ethapi/backend.go
#	miner/worker.go